[OWL-00034] Added stats from Delphi
