+++ src/twitch/commands/gamble.ts
@@ -7,6 +7,7 @@ import { isFeatureEnabled } from '@/lib/config';
 import { getCurrency, isNumber, weightedRandom } from '@/lib/utils';
 
 import { updateActivity } from '@/services/activity';
+import { saveGambleStats } from '@/services/stats';
 import { setTwitchUser } from '@/services/user';
 
 export const onGamble = async (
@@ -79,13 +80,7 @@ export const onGamble = async (
   await setTwitchUser(user.twitch_id!, { cash: newBalance });
 
   if (user.discord_id) {
-    await updateActivity(user.discord_id, {
-      $inc: {
-        [`gamble.${won ? 'total_wins' : 'total_losses'}`]: 1,
-        [`gamble.${won ? 'total_won' : 'total_lost'}`]: wager,
-      },
-      $set: { 'gamble.last_used': new Date() },
-      ...(won && { $max: { 'gamble.biggest_win': wager } }),
-    });
+    await saveGambleStats(user.discord_id, { won, wager });
+    await updateActivity(user.discord_id, { $set: { 'gamble.last_used': new Date() } });
   }
 };
